Updated the data type of cost from double to int.

+++ JavaAnalysis/src/AssignmentResult.java
@@ -1,9 +1,9 @@
 import java.util.*;
 public class AssignmentResult {
     public List<Integer> assignments;
-    public double totalCost;
+    public int totalCost;
 
-    public AssignmentResult(List<Integer> assignments, double totalCost) {
+    public AssignmentResult(List<Integer> assignments, int totalCost) {
         this.assignments = assignments;
         this.totalCost = totalCost;
     }
